Installer update

+++ installer.sh
@@ -4,15 +4,12 @@
 # not checked: echo
 dependencies="pwd grep printf sed egrep awk head tr cat curl"
 
-default_main_configuration_filepath="/etc/sendmail2mailgun/main.conf"
 executable_logging_name="sendmail2mailgun"
 sendmail2mailgun_path="./emulator.sh"
 
 # TODO
-# 1 - todos throughout the file
 # 2 - add check that sendmail2mailgun_path exists
 # future - add dry run mode
-# - Use local in functions 
 
 ##### Functions
 ### sanitize_variable_quotes
@@ -54,6 +51,7 @@ function load_cfg_file_variable()
 function check_dependencies()
 {
 	local error=0
+	local dependency
 	for dependency in $@; do
 		printf " - $dependency: "
 		if [ $(which $dependency) ]; then
@@ -76,7 +74,7 @@ function confirm()
 	printf "$1 | Type 'y' to confirm, anything else to reject: "
 	read -n 1 answer
 	echo ""
-	# shorthand for if [ "$answer" = "y" ] then; return 0; else return 1; fi 
+	# shorthand for if [ "$answer" = "y" ] then; return 0; else return 1; fi
 	[ "$answer" = "y" ]
 	return $?
 }
@@ -87,10 +85,11 @@ echo "**** $executable_logging_name installer ****"
 # parameter check
 has_config=1
 if [ -z "$1" ] || [ ! -f "$1" ]; then
-	echo "No path to a sendmail2mailgun configuration file (aka global configuration) provided as parameter. This sets the installer into configuration-file-less mode."
+	echo "No path to a $executable_logging_name configuration file (aka global configuration) provided as parameter."
+	echo "This sets the installer to configure $executable_logging_name in configuration-file-less mode."
 	echo "If this was not your intention, reject the confirmation below and re-run with the pattern"
-	echo "       ${0}  path/to/configuration/file "
-	echo "If you're new or simply want help creating a configuration, check out the template at ./configuration_templates/global.conf"
+	echo "       ${0}  path/to/configuration.file "
+	echo "A template may be found at ./configuration_templates/global.conf."
 	confirm "Confirm configuration-file-less installation? "
 	if [ $? -ne 0 ]; then
 		echo "Aborting..."
@@ -112,17 +111,17 @@ if [ $has_config -eq 1 ]; then
 		main_cfg_filepath="$1"
 	else
 		main_cfg_filepath="$(pwd)/$1"
-		echo "Main configuration absolute filepath: $main_cfg_filepath"
+		#echo "Global configuration absolute filepath: $main_cfg_filepath"
 	fi
-
-	# get configuration values from files, properly formatted without enclosing single or double quotes
+	echo "Global configuration $main_cfg_filepath used"
+	# get configuration values, properly formatted without enclosing single or double quotes
 	mailgun_api_account_configurations_folder=$(load_cfg_file_variable "$main_cfg_filepath" "mailgun_api_account_configurations_folder")
 	usecase_configurations_folder=$(load_cfg_file_variable "$main_cfg_filepath" "usecase_configurations_folder")
 	log_filepath=$(load_cfg_file_variable "$main_cfg_filepath" "log_filepath")
 fi
 
-echo "Installation:"
 if [ $has_config -eq 1 ]; then
+echo "Installation:"
 	printf " - Mailgun API account configurations folder: "
 	if [ ! -z "$mailgun_api_account_configurations_folder" ]; then
 		echo  "set to $mailgun_api_account_configurations_folder"
@@ -155,23 +154,17 @@ if [ $has_config -eq 1 ]; then
 	else
         	printf "configuration variable not set or empty. Logging will be disabled (can be overwritten on usecase basis and by the --log-filepath parameter)\n"
 	fi
-fi
 
-echo " - Executable <=> main configuration link:"
-final_configuration_filepath="$main_cfg_filepath"
-if [ $has_config -eq 1 ]; then
-	script_load_instruction_filepath="$default_main_configuration_filepath"
-else
-	script_load_instruction_filepath=""
-fi
-
-if [ ! -z "$script_load_instruction_filepath" ] && [ ! "$final_configuration_filepath" = "$script_load_instruction_filepath" ]; then
-	echo "   Currently, the configuration is not where $executable_logging_name expects it."
-	echo "   Expected: $default_main_configuration_filepath, Current: $main_cfg_filepath"
+	echo " - Executable <=> main configuration link:"
+	echo "   If you want this global configuration to be loaded by default, this installer can adapt $sendmail2mailgun_path to do so"
+	echo "   Do you want this and if yes, shall it stay at the current location $main_cfg_filepath or be moved?"
 	echo "   Options:"
-	echo "     (1) move configuration from current to expected location"
-	echo "     (2) keep the current location (adapts $executable_logging_name's \"configuration load\" instruction)"
-	echo "     (3) choose yet another path (combination of (1) and (2))"
+	echo "     ( ) set as default global configuration at ..."
+	echo "          (1) the current location"
+	echo "          (2) another location (file will be moved)"
+	echo "     (3) don't set as default global configuration"
+	echo "   Please be aware that option (3) sets $executable_logging_name in \"configuration-file less\" mode which requires the flag --cfg <filepath>"
+	echo "   to indicate a global configuration"
 	printf "   Answer: "
 	while true; do
 		read -n 1 -s answer
@@ -181,19 +174,18 @@ if [ ! -z "$script_load_instruction_filepath" ] && [ ! "$final_configuration_fil
 		fi
 		if [ "$answer" -eq 1 ]; then
 			echo 1
-			final_configuration_filepath="$script_load_instruction_filepath"
+			script_load_instruction_filepath="$main_cfg_filepath"
 			break
 		fi
 		if [ "$answer" -eq 2 ]; then
-			echo 2
-			script_load_instruction_filepath="$final_configuration_filepath"
-			break
-		fi
-		if [ "$answer" -eq 3 ]; then
-			printf "3 | Path: "
+			printf "2 | Path: "
 			read path
 			final_configuration_filepath="$path"
 			script_load_instruction_filepath="$path"
+			break
+		fi
+		if [ "$answer" -eq 3 ]; then
+			echo 3
 			break
 		fi
 	done
@@ -213,25 +205,32 @@ if [ ! -z "$final_configuration_filepath" ] && [ ! "$final_configuration_filepat
 		echo "OK"
 	fi
 	printf "   [Operation] Moving configuration file $main_cfg_filepath to $final_configuration_filepath: "
+	if [ -f "$final_configuration_filepath" ]; then
+		echo "Error"
+		echo "  $final_configuration_filepath exists, won't overwrite. Aborting..."
+		exit 1
+	fi
 	mv "$main_cfg_filepath" "$final_configuration_filepath"
 	if [ $? -ne 0 ]; then
 		echo "Error"
+		exit 1
         fi
 	echo "OK"
 fi
 
-if [ ! "$script_load_instruction_filepath" = "$default_main_configuration_filepath" ]; then
-	if [ -z "$script_load_instruction_filepath" ]; then
-		printf "   [Operation] Disabling $executable_logging_name \"configuration load\" instruction: "
-	else
-		printf "   [Operation] Adapting $executable_logging_name \"configuration load\" instruction to $script_load_instruction_filepath: "
-	fi
-	sed "s/^\s*configuration_filepath=\"\"/configuration_filepath=\"$script_load_instruction_filepath\"/" -i "$sendmail2mailgun_path"
-	# TODO sed change
-	# TODO check if successful
+if [ ! -z "$script_load_instruction_filepath" ]; then
+	printf "   [Operation] Adapting $executable_logging_name \"configuration load\" instruction to $script_load_instruction_filepath: "
+	sed  "s#^configuration_filepath=\"\"#configuration_filepath=\"$script_load_instruction_filepath\"#" -i "$sendmail2mailgun_path"
 	echo "OK"
 fi
 
 # TODO make executable
 # TODO protect configuration files
 echo "Installer finished"
+echo "IMPORTANT: protect the configuration/keyfile holding the Mailgun API keys with appropriate permissions - their access should be restricted to the user running $executable_logging_name"
+echo "You may also want to make $sendmail2mailgun_path executable and link/copy into somewhere into \$PATH($PATH) to be able to execute it globally"
+if [ $has_config -eq 0 ]; then
+	echo ""
+	echo "${executable_logging_name}'s flags are documented at https://github.com/DonTseTse/sendmail2mailgun/#flags"
+	echo "In configuration-file-less mode, --domain <domain> and --keyfile <keyfile>, are compulsory, make sure to check out https://github.com/DonTseTse/sendmail2mailgun/#configuration"
+fi
